refactor: api 수정에 따른 뷰 로직 수정
