+++ src/main/resources/templates/reservation.html
@@ -110,7 +110,6 @@
                     <div class="theme-card-body">
                         <div class="theme-card-head">
                             <strong th:text="${theme.name}">미드나잇 시그널</strong>
-                            <span th:text="${theme.durationTime}">01:30</span>
                         </div>
                         <p th:text="${theme.description}">설명</p>
                     </div>
@@ -134,7 +133,6 @@
                 <div class="theme-card-body">
                     <div class="theme-card-head">
                         <strong th:text="${theme.name}">인기 테마</strong>
-                        <span th:text="${theme.durationTime}">01:30</span>
                     </div>
                     <p th:text="${theme.description}">설명</p>
                 </div>
